+++ CHANGELOG.md
@@ -159,6 +159,9 @@ All notable changes to this project will be documented in this file. This change
 - Remove clj-time library
 - Change asset news database to use XTDB
 
+### 2023-09-23
+- Asset news bot message 
+
 ## [0.1.1] - 2023-03-25
 ### Changed